+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+- The error hierarchy: every Mistri failure rescues as `Mistri::Error`.
+
 ## [0.0.3] - 2026-07-04
 
 - Repository moved to github.com/mcheemaa/mistri.